# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hybrid firewall in Go using public blocklists",
        "title": "Meds: net healing",
        "contact": {
            "name": "cnaize",
            "url": "https://github.com/cnaize"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "v1.2.0"
    },
    "paths": {
...
        "/v1/domains/blocklist/exclude": {
            "get": {
                "description": "get all excluded domains",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Get excluded domains",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetDomainsResp"
                        }
                    }
                }
            },
            "post": {
                "description": "upsert domains to excludelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Upsert excluded domains",
                "parameters": [
                    {
                        "description": "domains to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertDomainsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "remove domains from excludelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Remove excluded domains",
                "parameters": [
                    {
                        "description": "domains to remove",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RemoveDomainsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/domains/blocklist/exclude/{domain}": {
            "get": {
                "description": "check if a domain is excluded",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Check excluded domains",
                "parameters": [
                    {
                        "type": "string",
                        "description": "domain to check",
                        "name": "domain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CheckDomainResp"
                        }
                    }
                }
            }
        },
        "/v1/domains/blocklist/include": {
            "get": {
                "description": "get all included domains",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Get included domains",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetDomainsResp"
                        }
                    }
                }
            },
            "post": {
                "description": "upsert domains to includelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Upsert included domains",
                "parameters": [
                    {
                        "description": "domains to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertDomainsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "remove domains from includelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Remove included domains",
                "parameters": [
                    {
                        "description": "domains to remove",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RemoveDomainsReq"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/domains/blocklist/include/{domain}": {
            "get": {
                "description": "check if a domain is included",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Domains"
                ],
                "summary": "Check included domains",
                "parameters": [
                    {
                        "type": "string",
                        "description": "domain to check",
                        "name": "domain",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CheckDomainResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.CheckDomainResp": {
            "type": "object",
            "properties": {
                "found": {
                    "type": "boolean"
                }
            }
        },
...
        "api.GetDomainsResp": {
            "type": "object",
            "properties": {
                "domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "good.com",
                        "bad.com"
                    ]
                }
            }
        },
...
        "api.RemoveDomainsReq": {
            "type": "object",
            "properties": {
                "domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "good.com",
                        "bad.com"
                    ]
                }
            }
        },
...
        "api.UpsertDomainsReq": {
            "type": "object",
            "properties": {
                "domains": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "good.com",
                        "bad.com"
                    ]
                }
            }
        },
...
    }
}